+++ ChangeLog.md
@@ -7,6 +7,7 @@ Note: This file only contains high level features or important fixes.
 * New combined compass and attitude instrument
 * You can select between multiple instruments by clicking on the instrument conntrol on a desktop build or press and hold on mobile builds
 * Support for Fly View and Joystick custom mavlink actions has changed. Both the name and formation of the command file is different now. Go to QGC docs to understand how it works now.
+* Support for setting individual Mavlink message rates in the Mavlink Inspector.
 
 ## 4.1